Single component for assistant list / Added sharing to assistant listing API
